tested clean build

+++ build_graph.bash
@@ -5,12 +5,14 @@ set -eux
 echo $GRAPH_DB 
 echo $EAC_DIR
 echo GRAPH_ML
-read -p "answer y to blow away the old index? " -n 1
-if [[ ! $REPLY =~ ^[Yy]$ ]]
-then
-    exit 1
+if [[ -e $GRAPH_DB ]]; then
+  read -p "answer y to blow away the old index? " -n 1
+  if [[ ! $REPLY =~ ^[Yy]$ ]]
+  then
+      exit 1
+  fi
+  rm -r $GRAPH_DB
 fi
-rm -r $GRAPH_DB
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" # http://stackoverflow.com/questions/59895
 export JAVA_OPTIONS="-Xms32m -Xmx5g"
 # create new database
